Fliter API and Drop Down APIs

+++ src/main/java/com/bookmyshow/main/repository/CityRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bookmyshow.main.model.City;
 
-public interface CityRepository extends JpaRepository<City, Long> {
+public interface CityRepository extends JpaRepository<City, Integer> {
 
 }
